+++ MapboxSearch/sample/src/androidTest/java/com/mapbox/search/sample/ExampleActivitiesTest.kt
@@ -2,6 +2,7 @@ package com.mapbox.search.sample
 
 import android.app.Activity
 import androidx.test.core.app.ActivityScenario
+import com.mapbox.search.sample.api.AddressAutofillKotlinExampleActivity
 import com.mapbox.search.sample.api.CategorySearchJavaExampleActivity
 import com.mapbox.search.sample.api.CategorySearchKotlinExampleActivity
 import com.mapbox.search.sample.api.CustomIndexableDataProviderJavaExample
@@ -20,7 +21,6 @@ import com.mapbox.search.sample.api.OfflineSearchJavaExampleActivity
 import com.mapbox.search.sample.api.OfflineSearchKotlinExampleActivity
 import com.mapbox.search.sample.api.ReverseGeocodingJavaExampleActivity
 import com.mapbox.search.sample.api.ReverseGeocodingKotlinExampleActivity
-import com.mapbox.search.sample.maps.MapsIntegrationExampleActivity
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
@@ -38,11 +38,9 @@ class ExampleActivitiesTest(private val clazz: Class<out Activity>) : MockServer
         @Parameterized.Parameters
         @JvmStatic
         fun testParameters(): Collection<Class<out Activity>> = listOf(
-            SimpleUiSearchActivity::class.java,
+            AddressAutofillKotlinExampleActivity::class.java,
 
             CustomThemeActivity::class.java,
-
-            MapsIntegrationExampleActivity::class.java,
 
             CustomIndexableDataProviderKotlinExample::class.java,
             CustomIndexableDataProviderJavaExample::class.java,