+++ MapboxSearch/ui/src/main/java/com/mapbox/search/ui/utils/ThemePatcher.kt
@@ -1,7 +1,6 @@
 package com.mapbox.search.ui.utils
 
 import android.content.Context
-import androidx.annotation.StyleRes
 import androidx.appcompat.view.ContextThemeWrapper
 import com.mapbox.search.ui.R
 import com.mapbox.search.ui.utils.extenstion.resolveAttr
@@ -14,25 +13,10 @@ private class SearchSdkContextThemeWrapper(context: Context) : ContextThemeWrapp
     )
 )
 
-private class SearchSdkPopupDialogContextThemeWrapper(
-    context: Context,
-    @StyleRes overlayTheme: Int
-) : ContextThemeWrapper(context, overlayTheme)
-
 internal fun wrapWithSearchTheme(context: Context): Context {
     return if (context is SearchSdkContextThemeWrapper) {
         context
     } else {
         SearchSdkContextThemeWrapper(context)
     }
-}
-
-internal fun wrapWithSearchPopupDialogThemeOverlay(context: Context): Context {
-    if (context !is SearchSdkPopupDialogContextThemeWrapper) {
-        val overlayTheme = context.resolveAttr(R.attr.mapboxSearchSdkPopupDialogThemeOverlay)
-        if (overlayTheme != null) {
-            return SearchSdkPopupDialogContextThemeWrapper(context, overlayTheme)
-        }
-    }
-    return context
 }