+++ MapboxSearch/ui/src/main/java/com/mapbox/search/ui/utils/extenstion/AsyncOperationTask.kt
@@ -1,7 +1,0 @@
-package com.mapbox.search.ui.utils.extenstion
-
-import com.mapbox.search.AsyncOperationTask
-
-@get:JvmSynthetic
-internal val AsyncOperationTask?.isCompleted: Boolean
-    get() = this != null && !(isDone || isCancelled)