+++ MapboxSearch/ui/src/main/java/com/mapbox/search/ui/view/category/CategoryViewCallback.kt
@@ -1,6 +1,0 @@
-package com.mapbox.search.ui.view.category
-
-internal interface CategoryViewCallback {
-
-    fun onItemClick(categoryEntry: CategoryEntry)
-}