+++ MapboxSearch/ui/src/test/java/com/mapbox/search/ui/DataClassGeneratedFunctionsTest.kt
@@ -8,11 +8,9 @@ import com.mapbox.search.common.tests.isEnum
 import com.mapbox.search.record.FavoriteRecord
 import com.mapbox.search.record.IndexableRecord
 import com.mapbox.search.ui.tests_support.UiCustomTypeObjectCreators
-import com.mapbox.search.ui.view.SearchBottomSheetView
 import com.mapbox.search.ui.view.SearchMode
 import com.mapbox.search.ui.view.category.Category
-import com.mapbox.search.ui.view.favorite.FavoriteTemplate
-import com.mapbox.search.ui.view.feedback.IncorrectSearchPlaceFeedback
+import com.mapbox.search.ui.view.place.IncorrectSearchPlaceFeedback
 import com.mapbox.search.ui.view.place.SearchPlace
 import com.mapbox.test.dsl.TestCase
 import nl.jqno.equalsverifier.EqualsVerifier
@@ -34,14 +32,12 @@ internal class DataClassGeneratedFunctionsTest {
         listOf(
             // regular classes
             SearchPlace::class,
-            FavoriteTemplate::class,
             Category::class,
             // TODO(#737): enable test for SearchResultFeedback
             // IncorrectSearchPlaceFeedback.SearchResultFeedback::class,
             IncorrectSearchPlaceFeedback.FavoriteFeedback::class,
             IncorrectSearchPlaceFeedback.HistoryFeedback::class,
             // enums
-            SearchBottomSheetView.CollapsedStateAnchor::class,
             SearchMode::class,
         )
             .forEach { clazz ->